python-fund-updater/requirements.txt
aiohttp
motor
python-dotenv